Fix Enemy Place Bug
There was a really weird bug. I was having a really difficult time figuring it out. But it was registering that there was a card selected but it kept saying there wasn't but the code that was saying there wasn't a card selected only ran if there was a card selected. There's probably some junk code in there now but I added in some redundancy and it works. No clue why it didn't before.