Add Blood Return and Spell
I added the blood return effect where if a player dies on their side of the board they gets 100% of the blood back. But if they dies on the enemy side they get back 50%.

I also added the fireball spell and possibly made it simpler to create spell cards by adding an abstract class.