Restructure Files
I restructured the files so hopefully now it should be easier to go through. I think here I also fixed a bug but that might have been a previous push. I was so out of it. Also refactored the way the attack phase works a little bit. There was a bug that didn't update the enemy blood level after we attack. This fixes that bug as well.