Implement Place Card Visual
I implemented the animation for cards for when you click on them and when you move them. Now the card slots will glow white whenever you select a card. I also added in the behavior of deselecting the card in your hand when you right click to move a card.