Update UI
I updated the UI for the card description and the turn state UI. So now we have a persistent interface that shows us what state we're in and how to get to the next state.

There was also a bug that just played the medium summon sound. That is now fixed so we should hear all the summon sounds now