Fix Bug
There was a bug where if you hovered over the enemy cards while they were holding it, you would be able to see what card it was. Not necessarily a bad thing but I don't think it makes sense to hide them visually and then show them in the overlay. So I disabled that.