Fix Bug
There was a bug that wasn't returning the correct blood amount back to the player and enemy. The player was always receiving full blood back and the enemy was always receiving half blood back. It is now fixed to where the player receives blood depending on where the card died and the enemy as well. Also a note, the play area should be able to scale dynamically to whatever values we want.

Also changed folder structure and got rid of some scenes