Add Mouse UI
I added in a small mouse indicator. So when you hover over things it contextually tells you what you can do. So far this is working well.

Only one known bug is when you have a card selected and you hover on the edge of a card slot that has a card occupied it says "unselect card" even though the actions are either to sacrifice or move the card. This only happens on the edge of the card slot so I don't think it's that big of a deal