Add Enemy Passive AI
I reworked how the base enemy class works a bit. Added a few functions and I also created the enemy passive AI. I created a new scene for it but I haven't changed the environment so visually it looks the same but the AI definitely reacts much differently.

Also fixed a bug that didn't register the right position when moving a card. It was delayed by one move. But now it works so the blood returnal system works as intended now. Fingers crossed.

Also I reworked the scene manager script so it's more flexible when moving between scenes but at the moment the flow isn't correct. Easily adjustable though. I can fix that in the next push