Add Enemy AI
I spent some time playtesting this new enemy AI and some edge cases came up. I added in the functionality for a stalemate to occur. So far I set up the event but nothing happens with it yet.

There was also a bug where when you ran out of cards the whole game would freak out. So I added in functionality when you ran out of cards to continue the game without drawing.

I refactored some of the code to remove a few get components so the game should be marginally more performant I hope.

The biggest thing is the new enemy AI. I called it smart, I think that might have been a bit optimistic in the name but it's the most complex AI in the game. I haven't lost to it. But it is definitely the most challenging and really aggressive.