Add Large Display and Fix bug
I added in a large display because during player testing I noticed that people were still having trouble knowing what they had to do. So I added in a UI display that triggers whenever the player has to complete an action to continue.

Also there was a bug that didn't reset the sacrifice cards when you unselected a card indirectly such as clicking a card to move. That is now fixed