Add Dynamic Cursor, Adjust Cursor UI, Fix Bug
I added in a dynamic cursor so when you hover over something you can interact with it will actually change the cursor it self.

I also adjusted the Cursor UI to take into account the actual size of the cursor. So now it won't be under the cursor no matter what size.

There was also a bug that didn't update the marked for sacrifice cards when you ended a turn. Now when you end a turn. The marked cards all unmark themselves.