Add Tooltip, Fix Bugs
I added in a tooltip for the player so they're not left completely lost on why things aren't working.

There was also a bug caused from a previous fix that would always return the the blood of the card even if it was sacrificed. This is now fixed.

Another bug was in the smart AI that didn't correctly pick a card. The problem was if the player was unable to continue playing because of their low blood level the enemy AI would play a card depending on whether there were cards in play or not. But if the cards in play were on the player home row, they would not register the enemy checking their own hand, causing a null error. This is now fixed.